+++ 2017-1/zkz/Tree3/Tree/test.c
@@ -20,11 +20,25 @@ int main() {
 	int size2 = 5;
 
 	/*初始化查找表*/
-	AVL avl = newAVL(8);
+	BST bst = newBST(inputs[0]);
+	AVL avl = newAVL(inputs[0]);
 	for (int i = 1; i < size1; ++i) {
+		bst = BSTinsert(bst, inputs[i]);
 		avl = AVLinsert(avl, inputs[i]);
 		if (DEBUG)SuperShow(avl),printf("\n");
 	}
+	printf("BST:\n");
+	Show(bst);
+	printf("\n");
+	/*开始查找操作*/
+	for (int j = 0; j < size2; ++j) {
+		/*如果找到就删除，没找到就插入*/
+		BSTfind(bst, toFind[j]) ? BSTremove(bst, toFind[j]) : BSTinsert(bst, toFind[j]);
+		Show(bst);
+		printf("\n");
+	}
+
+	printf("AVL:\n");
 	Show(avl);
 	printf("\n");
 	/*开始查找操作*/
